Improve proxy commands
Split each command in a separate file.
Validate the specified port.
Introduce constants for http status codes.
Remove USE_PROXY - it does not make any sense as it cannot be set to false.
