+++ bootstrap.ts
@@ -108,9 +108,10 @@ $injector.require("credentialsService", "./services/credentials-service");
 $injector.requireCommand("dev-preuninstall", "./commands/preuninstall");
 $injector.requireCommand("dev-generate-messages", "./commands/generate-messages");
 $injector.requireCommand("doctor", "./commands/doctor");
-$injector.requireCommand("proxy|*get", "./commands/proxy");
-$injector.requireCommand("proxy|set", "./commands/proxy");
-$injector.requireCommand("proxy|clear", "./commands/proxy");
+
+$injector.requireCommand("proxy|*get", "./commands/proxy/proxy-get-command");
+$injector.requireCommand("proxy|set", "./commands/proxy/proxy-set-command");
+$injector.requireCommand("proxy|clear", "./commands/proxy/proxy-clear-command");
 
 $injector.require("utils", "./utils");
 $injector.require("bplistParser", "./bplist-parser");
