+++ commands/proxy/proxy-clear-command.ts
@@ -1,0 +1,18 @@
+import { ProxyCommandBase } from "./proxy-base-command";
+const proxyClearCommandName = "proxy|clear";
+
+export class ProxyClearCommand extends ProxyCommandBase {
+	constructor(protected $analyticsService: IAnalyticsService,
+		protected $logger: ILogger,
+		protected $proxyService: IProxyService) {
+		super($analyticsService, $logger, $proxyService, proxyClearCommandName);
+	}
+
+	public async execute(args: string[]): Promise<void> {
+		this.$proxyService.clearCache();
+		this.$logger.out("Successfully cleared proxy.");
+		await this.tryTrackUsage();
+	}
+}
+
+$injector.registerCommand(proxyClearCommandName, ProxyClearCommand);
